+++ .scripts/update_version.sh
@@ -10,14 +10,6 @@ fi
 
 echo "Setting the version to: $version"
 
-for package in packages/client-node packages/client-web; do
-  if [ -f "$package/package.json" ]; then
-    echo "Updating client-common version in $package/package.json"
-    json=$(cat "$package/package.json")
-    echo "$json" | jq --arg version "$version" '.dependencies["@clickhouse/client-common"] = $version' > "$package/package.json"
-  fi
-done
-
 for package in packages/client-common packages/client-node packages/client-web; do
   if [ -f "$package/package.json" ]; then
     echo "Updating version in $package/src/version.ts"